# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Task API
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development servers
  - url: https://task-api-uaop.onrender.com
    description: Production server
tags:
  - name: Tasks
    description: Task-related endpoints
  - name: Monitoring
    description: Operational endpoints, such as health checks
paths:
  /tasks:
    get:
      summary: List all tasks
      tags:
        - Tasks
      parameters:
        - name: completed
          in: query
          required: false
          description: Filter tasks by completion status
          schema:
            type: string
            enum: ['true', 'false']
          examples:
            completed_true:
              value: 'true'
              description: Returns only completed tasks
            completed_false:
              value: 'false'
              description: Returns only incomplete tasks
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '400':
          description: Bad request due to invalid query parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
    post:
      summary: Create a new task
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request due to invalid data
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
                example:
                  errors:
                    - 'Title must be at least 3 characters'
  /health:
    get:
      summary: Health Check
      description: Returns the current status of the API. Used by Render to verify that the service is running.
      tags:
        - Monitoring
      responses:
        '200':
          description: API is running successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          readOnly: true
          type: integer
          example: 1
        title:
          type: string
          example: Buy groceries
        completed:
          type: boolean
          example: false